Create main.py
 In this python program we see how a Flask application interacts with SQLite
Make sure that the SQLite database ‘database.db’ and items’ table are ready for this web application otherwise create one by uncommenting the Create query.

Our Flask application has three View functions.
First new_item() function is bound to the URL rule (‘/enternew’). It renders an HTML file containing item information form.
Entry form data is posted to the ‘/addrec’ URL which binds the addrec() function.

This addrec() function retrieves the form’s data by POST method and inserts in  items table. Message corresponding to success or error in insert operation is rendered to ‘report.html’.

The application contains another list() function represented by ‘/list’ URL. It populates ‘rows’ as a MultiDict object containing all records in the items table. This object is passed to the list.html template.

This list.html is a template, which iterates over the row set and renders the data in an HTML table.

Finally, the ‘/’ URL rule renders a ‘index.html’ which acts as the entry point of the application.